@@ -1,32 +1,35 @@
 {
   "name": "3d-game-of-life",
-  "version": "1.8.1",
+  "version": "1.9.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/SamuelTribeUK/3D-Game-of-Life.git"
   },
   "homepage": "https://github.com/SamuelTribeUK/3D-Game-of-Life",
   "dependencies": {
-    "express": "~4.17.1",
+    "3d-game-of-life": "file:",
+    "chart.js": "^4.4.8",
+    "express": "^4.21.2",
+    "node": "^22.9.0",
     "three": "~0.125.2"
   },
   "devDependencies": {
     "@babel/preset-env": "~7.13.10",
-    "babel-jest": "~26.6.3",
+    "babel-jest": "^29.7.0",
     "clean-webpack-plugin": "~3.0.0",
     "css-loader": "~5.0.1",
     "file-loader": "~6.2.0",
     "html-webpack-plugin": "~5.0.0",
-    "jest": "~26.6.3",
+    "jest": "^29.7.0",
     "jquery": "~3.5.1",
-    "jsdoc": "~3.6.6",
+    "jsdoc": "^4.0.4",
     "style-loader": "~2.0.0",
     "three": "~0.125.2",
     "toastify-js": "~1.9.3",
-    "webpack": "~5.21.2",
+    "webpack": "^5.98.0",
     "webpack-bundle-analyzer": "~4.4.0",
-    "webpack-cli": "~4.5.0",
-    "webpack-dev-server": "~3.11.2",
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.0",
     "webpack-merge": "~5.7.3",
     "worker-loader": "~3.0.8",
     "workerloader-jest-transformer": "0.0.3"
